refactor(CodeEditor): reorganize testing utilities
- Moved `codeSnippets` to a dedicated testing file for better organization and accessibility.
- Updated import paths in the CodeEditor stories and specs to reflect the new location of `codeSnippets`.
- Introduced comprehensive test utilities for CodeMirror extension hooks, enhancing the testing framework's capabilities.
