+++ packages/code-editor/src/CodeEditor/CodeEditor.spec.tsx
@@ -2,8 +2,9 @@ import { forceParsing } from '@codemirror/language';
 import { EditorState } from '@codemirror/state';
 import { act, waitFor } from '@testing-library/react';
 
+import { renderCodeEditor } from '../testing/testUtils';
+
 import { LanguageName } from './hooks/extensions/useLanguageExtension';
-import { renderCodeEditor } from './testing/testUtils';
 import { CopyButtonAppearance } from './CodeEditor.types';
 import { CodeEditorSelectors } from '.';
