+++ packages/code-editor/src/testing/codeSnippets.ts
